rename costing pkg
